fix: fix bug in dropdown with initial transition between menus
Issue: initial transition between menus didn't animate height

Cause of Issue: The menu is transitioning height, and when the initial
value of height is auto, it won't know how to transition from auto ->
next menu height.

Solution: When the dropdown menu is mounted, set the list height to be
its height (with the initial menu). Thus, it can transition immediately
to another height smoothly.