feat: begin working on animating mode (read more) fix: fix traversal bug with empty leaves
-- feat --
Added animation context to more easily access the isAnimateMode state
to change the UI based on isAnimateMode.

-- fix --
The traversal algorithms forgot to take into account the fact that we
deliberately set empty children in a child pair to still exist in the
tree but have no name or uuid. Changed code to take account that.