fix: fix bug with positioning for animate element (Read more), fix bug with ref not updating functions;
-- fix 1 --
Bug --
When the user moves the canvas, then animates, the animated element
would start in the wrong place (b/c its initial props never change).

Solution --
Add a ref that keeps track of the displacement of the element, and apply
it directly to the element's transform. Thus, it will always be in sync
with the position of the canvas.

-- fix 2 --
Bug --
With the handleRun function in useAnimationControl, it didn't seem to
update when the animationState changed. Thus its closure would contain
variables that were stale, like animationState.

Solution --
It turns out that because I was destructuring the ref in another file,
that ref automatically became stale, b/c now we can't access ref.current
directly, which holds the new value. Solution was to not destructure the ref to ensure that it's
actually updated.