refactor: update DropdownSelect to be less redundant
Changes:
Instead of a render props approach, we pass in a list of values instead.
These values will contain the props for a SelectableDropdownItem as this
is assumed to be the only possible item for a DropdpownSelect. In the
future, we may refactor this to be more flexible, working with any
custom component, so it could look like a list of components with props
already included.

Reasoning:
The render props approach was too redundant as there were now two
separate states for the same purpose (selectItem and whatever state we
needed to keep track of in the parent)