fix: fix more issues with stepping forward and backward (read more)
-- Stepping Backward --
Accounted for cases where the animation element is in transit, and
specifically for the case of step backwards, we need to check whether
it's in transit to see whether to go back to its original node or to the
previous node.

-- Stepping Forward --
Simplified implementation. Doesn't need to rely on animating state b/c
we came to the realization that we will always increment current step by 1 regardless of whether it's in
transit.