feat: more functionality for control tooltips (read more)
1. added tooltips for disabled step back and forward buttons
2. disabled skip to end when already at the end based on animation
state.

-- issues --
Currently I'm trying to disabled the step buttons and reset button based
on where the animation is currently at (i.e. at the beginning, reset and
step back should be disabled) Main issue is we don't have access to the
currentStep. Do we expose this in the AnimationContext and risk
re-rendering every time a step is updated?