refactor: remove collapse/expand animation for actions (read more)
-- Reasoning --
The animation relied heavily on overflow hidden to hide the section
contents as they faded upwards. The main issue with this is that now,
any item that we want to absolutely position in the actual section
content, will get clipped iff their parent is positioned
relatively. This is usually the case for the absolutely positioned
elements like dropdown. We want it to be relative to their containing
block.

If this change wasn't made, we'd be spending way too much time tinkering
with React portals to get the dropdowns to show up properly and work
with scroll. Thus, if we ever get to it, we'll come back to this.