fix: fix stutter issue with animation when new steps generated
-- Issue --
Whenever new steps were generated, animation would sometimes stutter.
This was because the useEffect beginning the animation was called twice
due to dependencies on both isAnimatingMode and steps (where steps were
generated after isAnimatingMode was affected, thus the two updates).

-- Solution --
Make sure in the useEffect with isAnimatingMode and steps as
dependencies, we wait for steps to be updated before running animation.
We simply check if updateStepsRef.current is false, indicating that we
had just consumed the last step update and are ready to begin the
animation.