refactor: add intermediate animation component to handle initial props
-- Issue --
I couldn't update the 'from' prop for useSpring dynamically, so I
wouldn't receive any animation values for my new animation elements like
the text elements.

-- Solution --
I ensured that the props would be available on mount by creating an
intermediate component that handles setting the steps and elements,
while the actual animation containing the spring logic is now not
rendered. So as a bonus, in addition to solving this issue with the
'from' prop, we now get rid of animation spring logic when
we don't need it (i.e. when we're not animating).