fix: fix issue in animation where remounting the animation (read more)
-- issue --
Whenever we remounted the AnimationManagerInnerTest component, somehow
steps were skipping as we were animating. The main issue was that our
animationMethodsRef wasn't being changed, so previous state values were
kept, and messed with logic, causing the skipping of steps.

-- solution --
On unmount (via useEffect cleanup function), reset animationMethodsRef
to null such that on the next mount, we can render with updated values.