fix: animation issue with updated tree (read more)
-- issue --
Whenever we ran an animation, exited animating mode, then updated the
tree, then re-entered animating mode, we always started off with the
previous animation's initial state. The issue here was that we weren't
refreshing the animation elements and steps prior to remounting
AnimationManagerInnerTest. This became a problem b/c
AnimationManagerInnerTest renders when isAnimatingMode is toggled, and
thus for a split second, before steps and elements are updated (due to
the tree updating), we have the previous animation's values.

-- solution --
On toggling isAnimatingMode off, remember to reset animationElements and
steps to null!