fix: fix issue with previous fix (read more)
-- issue --
With the last fix brought about a new problem where, b/c we reset steps
and animation elements every time we toggle isAnimatingMode off, we mess
up the case where we reanimate with the same props, b/c now they're
deleted.

-- solution --
Instead of deliberately resetting steps and animationElements, we now
just make sure that we have completely updated steps prior to actually
mounting the AnimationManagerInnerTest component. This works even with
the ref value b/c on the first mount when it checks isAnimatingMode, it
also now checks if we've updated steps. On the second mount, once we've
actually set the steps and elements and toggled off updateStepsRef,
our component will rerender (due to steps and elements being set) with this new ref value.