refactor: abstract d3 representation update into D3Context (read more)
-- issue --
We found situations where we needed to update the d3 representation from
anywhere (i.e. handleInsert for binary search tree), and we didn't want to have to go throwing around the visualization's width and height everywhere, as it wouldn't make sense in every case if we're looking at one function we need it in.

-- solution --
Abstract the updating of D3 representation of our data structure into
the D3 context. All we have to do on each page is pass in an updater
function that takes in our own representation of the structure, and
make sure to set the visualizationRef now in D3 context to our
visualization component (automatically done with DSAPageLayout so no
worries there.)