refactor: updated tree traversal to use React tree rendering (read more)
Migrated only the rendering portion of tree traversal to React from D3.

Still need to work on animation and ensuring that we're no longer
relying on d3 representation, but rather, our own.

In terms of what I learned in migrating from D3 rendering to React for
our data structures, it was that D3 is very much imperative in the way
it renders. We tell it 'hey draw these specific elements on the screen
exactly the way we want it positioned', while with React, it's more like
"hey here's what should be rendered on the screen based on our current
state" and when such state changes, so does the render. We're still
changing values in both cases, but how we get the screen to rerender is
different.

Some architectural stuff I ran into mainly dealt with creating a
'BaseTree' render props component, b/c we are going to have trees that
are very similar in that they need links and nodes, but
behavior-wise are different. In this case, I needed a tree that applied
click handlers to its node. The logic that I needed to abstract was
generating the links based on the nodes, so that's what happens in
BaseTree.